1. SapphireXR_App Manual Batch: flow controller 소수점 입력 기능 추가, Device Name Column 값 출력 기능 추가

+++ SapphireXR_App/Models/PLCService/PLCService.Write.cs
@@ -313,9 +313,9 @@ namespace SapphireXR_App.Models
             WriteFirstInterlockSetting(true, 1);
         }
 
-        public static void WriteFlowControllerTargetValue((string, int)[] aControllerIdTargetValues, short rampTime)
+        public static void WriteFlowControllerTargetValue((string, float)[] aControllerIdTargetValues, short rampTime)
         {
-            foreach((string id, int targetValue) in aControllerIdTargetValues)
+            foreach((string id, float targetValue) in aControllerIdTargetValues)
             {
                 WriteFlowControllerTargetValue(Util.RecipeFlowControlFieldToControllerID[id], targetValue, rampTime);
             }
